fix(helpers): fixed type in count_valid_subarray_sum_of_k

+++ src/helpers/count_valid_subarray_sum_of_k.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 pub fn count_valid_subarray_sum_of_k(
     logs: &[i32],
     k: i32,
-    min_len: i32,
+    min_len: isize,
 ) -> i32 {
     
     let mut map: HashMap<i32, Vec<isize>> = HashMap::new();
@@ -21,7 +21,7 @@ pub fn count_valid_subarray_sum_of_k(
         if let Some(indices) = map.get(&(sum - k)) {
             for &j in indices {
                 // (i -j + 1) >= min_len as isize or
-                if i - j >= min_len as isize - 1 {
+                if i - j >= min_len - 1 {
                     count += 1;
                 }
             }
